test: Extract delayNonBlocking to be reusable
