+++ Tests/SentryTests/Integrations/SentrySessionGeneratorTests.swift
@@ -113,18 +113,6 @@ class SentrySessionGeneratorTests: XCTestCase {
         autoSessionTrackingIntegration.install(with: options)
     }
     
-    private func delayNonBlocking(timeout: Double = 0.2) {
-        let group = DispatchGroup()
-        group.enter()
-        let queue = DispatchQueue(label: "delay", qos: .background, attributes: [])
-        
-        queue.asyncAfter(deadline: .now() + timeout) {
-            group.leave()
-        }
-        
-        group.wait()
-    }
-    
     private func goToForeground(forSeconds: TimeInterval = 0.2) {
         TestNotificationCenter.willEnterForeground()
         TestNotificationCenter.didBecomeActive()
